NSA-5-added functions for suspicious activities
Added functions to detect large packet size, to detect traffic to/from known malicious IP addresses to detect traffic on unusual ports to track and detect high-frequency traffic from a single source IP and to detect frequent connections to the same host (beaconing behavior). Also restructured the code and renamed the file
